Fix time base API (#1848)

+++ src/HAL/Include/nanoHAL_Time.h
@@ -12,27 +12,27 @@
 #include <nanoWeak.h>
 
 /// Our time origin is 1/1/1601 00:00:00.000.000.  In Gregorian Calendar Jan 1, 1601 was also a Monday.
-#define BASE_YEAR                   1601
-#define BASE_YEAR_LEAPYEAR_ADJUST   388    
-#define DAYS_IN_NORMAL_YEAR         365
-#define BASE_YEAR_DAYOFWEEK_SHIFT   1 
+#define BASE_YEAR                 1601
+#define BASE_YEAR_LEAPYEAR_ADJUST 388
+#define DAYS_IN_NORMAL_YEAR       365
+#define BASE_YEAR_DAYOFWEEK_SHIFT 1
 
-#define TIME_CONVERSION__TO_MILLISECONDS    10000
-#define TIME_CONVERSION__TO_SECONDS         10000000
-#define TIME_CONVERSION__TICKUNITS         10000
-#define TIME_CONVERSION__ONESECOND         1
-#define TIME_CONVERSION__ONEMINUTE         60
-#define TIME_CONVERSION__ONEHOUR           3600
-#define TIME_CONVERSION__ONEDAY            86400
+#define TIME_CONVERSION__TO_MILLISECONDS 10000
+#define TIME_CONVERSION__TO_SECONDS      10000000
+#define TIME_CONVERSION__TICKUNITS       10000
+#define TIME_CONVERSION__ONESECOND       1
+#define TIME_CONVERSION__ONEMINUTE       60
+#define TIME_CONVERSION__ONEHOUR         3600
+#define TIME_CONVERSION__ONEDAY          86400
 
-#define TIMEOUT_ZERO      LONGLONGCONSTANT(0x0000000000000000)
-#define TIMEOUT_INFINITE  LONGLONGCONSTANT(0x7FFFFFFFFFFFFFFF)
+#define TIMEOUT_ZERO     LONGLONGCONSTANT(0x0000000000000000)
+#define TIMEOUT_INFINITE LONGLONGCONSTANT(0x7FFFFFFFFFFFFFFF)
 
 // UNDONE: FIXME #define TIME_ZONE_OFFSET    ((signed long long)Time_GetTimeZoneOffset() * 600000000)
 
-#define NANOHAL_TIME_CONVERSION_MICRO_TO_SECONDS                1000000
+#define NANOHAL_TIME_CONVERSION_MICRO_TO_SECONDS 1000000
 
-#define TIME_UNIX_EPOCH_AS_TICKS           116444736000000000
+#define TIME_UNIX_EPOCH_AS_TICKS 116444736000000000
 
 /// NOTES: Why origin is at 1/1/1601.
 /// Current civil calendar is named as Gregorian calendar after Pope Gregory XIII as he made adjustments
@@ -44,64 +44,64 @@
 /// dates like 10/6/1582 does exist in history (see wiki), while CLR managed date/time will not throw an exception
 /// if you are to create that date. To stay safe side 1/1/1601 is taken as origin, as was done for Windows.
 
-
-uint64_t HAL_Time_SysTicksToTime(unsigned int sysTicks);
+uint64_t HAL_Time_SysTicksToTime(uint64_t sysTicks);
 
 /// <summary>
 /// System time and date for DateTime managed class.
-/// This value will be provided by the system tick or from an RTC if hardware support exists and if the board designer has enabled it in the configuration options.
-/// The datePartOnly allows returning only the date part with the time fields zeroed.
+/// This value will be provided by the system tick or from an RTC if hardware support exists and if the board designer
+/// has enabled it in the configuration options. The datePartOnly allows returning only the date part with the time
+/// fields zeroed.
 /// </summary>
 /// <returns>Returns current time in 100ns elapsed since 1/1/1601:00:00:00.000 UTC.</returns>
-uint64_t  HAL_Time_CurrentDateTime(bool datePartOnly);
+uint64_t HAL_Time_CurrentDateTime(bool datePartOnly);
 
 /// <summary>
-/// Time according to this system. 
+/// Time according to this system.
 /// </summary>
 /// <returns>Returns current time in 100ns elapsed since 1/1/1601:00:00:00.000 UTC.</returns>
-uint64_t  HAL_Time_CurrentTime();
+uint64_t HAL_Time_CurrentTime();
 
 /// <summary>
 /// Set UTC time of the system. This will be effective immediately.
 /// </summary>
 /// <param name="utcTime">In 100ns since 1/1/1601:00:00:00.000</param>
-void      HAL_Time_SetUtcTime(uint64_t utcTime);
+void HAL_Time_SetUtcTime(uint64_t utcTime);
 
 /// <summary>
 /// Retrieves time since device was booted.
 /// </summary>
 /// <returns>Time in 100ns.</returns>
-//signed long long HAL_Time_GetMachineTime();
+// signed long long HAL_Time_GetMachineTime();
 
 /// <summary>
-/// Converts 64bit time value to SystemTime structure. 64bit time is assumed as an offset from 1/1/1601:00:00:00.000 in 100ns.
+/// Converts 64bit time value to SystemTime structure. 64bit time is assumed as an offset from 1/1/1601:00:00:00.000 in
+/// 100ns.
 /// </summary>
 /// <returns>True if conversion is successful.</returns>
-bool HAL_Time_ToSystemTime(uint64_t time, SYSTEMTIME* systemTime);
+bool HAL_Time_ToSystemTime(uint64_t time, SYSTEMTIME *systemTime);
 
 /// <summary>
 /// Retrieves number of days given a month and a year. Calculates for leap years.
 /// </summary>
 /// <returns>S_OK if successful.</returns>
-HRESULT HAL_Time_DaysInMonth(signed int year, signed int month, signed int* days);
+HRESULT HAL_Time_DaysInMonth(signed int year, signed int month, signed int *days);
 
 /// <summary>
 /// Retrieves number of days since the beginning of the year given a month and a year. Calculates for leap years.
 /// </summary>
 /// <returns>S_OK if successful.</returns>
-HRESULT HAL_Time_AccDaysInMonth(signed int year, signed int month, signed int* days);
+HRESULT HAL_Time_AccDaysInMonth(signed int year, signed int month, signed int *days);
 
 /// <summary>
 /// Converts SYSTEMTIME structure to 64bit time, which is assumed as an offset from 1/1/1601:00:00:00.000 in 100ns.
 /// </summary>
 /// <returns>Time value.</returns>
-uint64_t HAL_Time_ConvertFromSystemTime(const SYSTEMTIME* systemTime);
+uint64_t HAL_Time_ConvertFromSystemTime(const SYSTEMTIME *systemTime);
 
 /// APIs to convert between types
-bool          HAL_Time_TimeSpanToStringEx( const int64_t& ticks, char*& buf, size_t& len );
-const char*   HAL_Time_CurrentDateTimeToString();
+bool HAL_Time_TimeSpanToStringEx(const int64_t &ticks, char *&buf, size_t &len);
+const char *HAL_Time_CurrentDateTimeToString();
 
-
-uint64_t  CPU_MillisecondsToTicks       ( uint64_t ticks  );
+uint64_t CPU_MillisecondsToTicks(uint64_t ticks);
 
 #endif //_NANOHAL_TIME_H_
