+++ targets/FreeRTOS_ESP32/ESP32_WROOM_32/common/targetHAL_Time.cpp
@@ -10,15 +10,17 @@
 #include <esp32_os.h>
 
 // Converts FreeRTOS Tickcount to .NET ticks (100 nanoseconds)
-uint64_t HAL_Time_SysTicksToTime(unsigned int sysTicks) 
+uint64_t HAL_Time_SysTicksToTime(uint64_t sysTicks)
 {
-    return (((int64_t)sysTicks * (int64_t)1000000 + (int64_t)configTICK_RATE_HZ - 1) / (int64_t)configTICK_RATE_HZ) * 10;
+    return (((int64_t)sysTicks * (int64_t)1000000 + (int64_t)configTICK_RATE_HZ - 1) / (int64_t)configTICK_RATE_HZ) *
+           10;
 }
 
 // because HAL_Time_SysTicksToTime needs to be called from C we need a proxy to allow it to be called from 'C' code
-extern "C" {
-    
-    uint64_t HAL_Time_SysTicksToTime_C(unsigned int sysTicks)
+extern "C"
+{
+
+    uint64_t HAL_Time_SysTicksToTime_C(uint64_t sysTicks)
     {
         return HAL_Time_SysTicksToTime(sysTicks);
     }
