+++ targets/TI-SimpleLink/common/platformHAL_Time.cpp
@@ -8,15 +8,16 @@
 #include <ti/sysbios/knl/Clock.h>
 
 // Converts TI RTOS clock tick period to .NET ticks (100 nanoseconds)
-uint64_t HAL_Time_SysTicksToTime(unsigned int sysTicks) 
+uint64_t HAL_Time_SysTicksToTime(uint64_t sysTicks)
 {
     return ((int64_t)sysTicks * (int64_t)Clock_tickPeriod) * 10;
 }
 
 // because HAL_Time_SysTicksToTime needs to be called from C we need a proxy to allow it to be called from 'C' code
-extern "C" {
-    
-    uint64_t HAL_Time_SysTicksToTime_C(unsigned int sysTicks)
+extern "C"
+{
+
+    uint64_t HAL_Time_SysTicksToTime_C(uint64_t sysTicks)
     {
         return HAL_Time_SysTicksToTime(sysTicks);
     }
